Move flags package
