+++ internal/flags/simulator.go
